Improves the performance of CoveragePolygonValidator and CoverageValidator by adding logic to avoid testing sections of ring vertices if the entire section does not overlap the target adjacent polygon.

Port of https://github.com/locationtech/jts/pull/1053

### World Polygons G99_5
Statistics: 323 geometries, 329,373 polygons, 459,540 vertices
Old code: 1579 ms
New code: 400 ms

### World Polygons G65
Statistics: 323 geometries, 319,646 polygons, 27,706,942 vertices
Old code: 5308 s
New code: 59 s